+++ sdk/instrumentation/instrumentation-opentelemetry/package.json
@@ -40,7 +40,7 @@
     "format": "prettier --write --config ../../../.prettierrc.json --ignore-path ../../../.prettierignore \"src/**/*.ts\" \"test/**/*.ts\" \"samples-dev/**/*.ts\" \"*.{js,json}\"",
     "generate:client": "autorest --typescript ./swagger/README.md",
     "integration-test:browser": "karma start --single-run",
-    "integration-test:node": "echo Skipped. #nyc mocha -r esm --require source-map-support/register --reporter ../../../common/tools/mocha-multi-reporter.js --timeout 5000000 --full-trace \"dist-esm/test/{,!(browser)/**/}/*.spec.js\"",
+    "integration-test:node": "nyc mocha -r esm --require source-map-support/register --reporter ../../../common/tools/mocha-multi-reporter.js --timeout 5000000 --full-trace \"dist-esm/test/{,!(browser)/**/}/*.spec.js\"",
     "integration-test": "npm run integration-test:node && npm run integration-test:browser",
     "lint:fix": "eslint package.json api-extractor.json src test --ext .ts --fix --fix-type [problem,suggestion]",
     "lint": "eslint package.json api-extractor.json src test --ext .ts",
@@ -81,19 +81,21 @@
     "@azure/core-auth": "^1.3.0",
     "@azure/core-tracing": "1.0.0-preview.14",
     "@azure/logger": "^1.0.0",
-    "tslib": "^2.2.0",
-    "@opentelemetry/api": "^1.0.3"
+    "@opentelemetry/api": "^1.0.3",
+    "@opentelemetry/instrumentation": "^0.27.0",
+    "tslib": "^2.2.0"
   },
   "devDependencies": {
+    "@azure-tools/test-recorder": "^1.0.0",
     "@azure/dev-tool": "^1.0.0",
     "@azure/eslint-plugin-azure-sdk": "^3.0.0",
     "@azure/identity": "^2.0.1",
-    "@azure-tools/test-recorder": "^1.0.0",
     "@microsoft/api-extractor": "^7.18.11",
     "@types/chai": "^4.1.6",
     "@types/chai-as-promised": "^7.1.0",
     "@types/mocha": "^7.0.2",
     "@types/node": "^12.0.0",
+    "@types/sinon": "^10.0.6",
     "chai": "^4.2.0",
     "chai-as-promised": "^7.1.1",
     "cross-env": "^7.0.2",
@@ -120,12 +122,11 @@
     "prettier": "^1.16.4",
     "rimraf": "^3.0.0",
     "rollup": "^1.16.3",
+    "sinon": "^12.0.1",
     "source-map-support": "^0.5.9",
     "typedoc": "0.15.2",
     "typescript": "~4.2.0",
-    "util": "^0.12.1",
-    "sinon": "^12.0.1",
-    "@types/sinon": "^10.0.6"
+    "util": "^0.12.1"
   },
   "//sampleConfiguration": {
     "skipFolder": false,
