--- sdk/instrumentation/instrumentation-opentelemetry/sample.env
@@ -1,7 +1,0 @@
-# Used in most samples. Retrieve these values from an instance in the Azure
-# Portal. The APPCONFIG_TEST_SETTING_KEY value indicates which key to use
-# when retrieving an example setting.
-
-APPCONFIG_ENDPOINT="https://<resource name>.azconfig.io",
-APPCONFIG_TEST_SETTING_KEY="<key>"
-
