+++ sdk/instrumentation/instrumentation-opentelemetry/src/constants.ts
@@ -2,3 +2,6 @@
 // Licensed under the MIT license.
 
 export const SDK_VERSION: string = "1.0.0-beta.1";
+
+// https://www.w3.org/TR/trace-context/#version
+export const TRACEPARENT_HEADER_VERSION = "00";
