+++ sdk/instrumentation/instrumentation-opentelemetry/src/index.ts
@@ -1,4 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
+
 export * from "./logger";
 export * from "./instrumenter";
+export * from "./instrumentation";
