+++ sdk/instrumentation/instrumentation-opentelemetry/src/logger.ts
@@ -6,6 +6,5 @@ import { createClientLogger } from "@azure/logger";
 /**
  * The \@azure/logger configuration for this package.
  *
- * @internal
  */
 export const logger = createClientLogger("instrumentation-opentelemetry");
