--- sdk/instrumentation/instrumentation-opentelemetry/test.env
@@ -1,26 +1,0 @@
-# Used in most samples. Retrieve these values from an instance in the Azure
-# Portal.
-
-APPCONFIG_ENDPOINT="https://<resource name>.azconfig.io",
-
-# Used in the tests to indicate the key to use when retrieving a setting from
-# the App Configuration and the expected value of that setting.
-
-APPCONFIG_TEST_SETTING_KEY="<key>"
-APPCONFIG_TEST_SETTING_EXPECTED_VALUE="<expected value>"
-
-# Used to authenticate using Azure AD as a service principal for role-based
-# authentication in the tokenAuth sample.
-#
-# See the documentation for `EnvironmentCredential` at the following link:
-# https://docs.microsoft.com/javascript/api/@azure/identity/environmentcredential
-
-AZURE_TENANT_ID="<AD tenant id or name>"
-AZURE_CLIENT_ID="<ID of the user/service principal to authenticate as>"
-AZURE_CLIENT_SECRET="<client secret used to authenticate to Azure AD>"
-
-# Our tests assume that TEST_MODE is "playback" by default. You can change it
-# to "record" to generate new recordings, or "live" to bypass the recorder
-# entirely.
-
-# TEST_MODE="playback"
